refactor(map-calibration): delete MapRectLocator NCC ladder; relocate MapRect to its own file
