+++ tests/Mithril.MapCalibration.Tests/Detection/ReplayFixtureTests.cs
@@ -121,8 +121,8 @@ public sealed class ReplayFixtureTests
         var tex = WicImageLoader.LoadGray(texturePath);
 
         // The screenshot here is assumed already cropped to the map rect (1:1
-        // texture extent) for the replay; real captures crop via MapRectLocator
-        // in Phase 2.
+        // texture extent) for the replay; real captures crop via the production
+        // IMapRegionRefiner (FeatureMatchingRefiner) in Phase 2.
         var rect = new MapRect(0, 0, shot.Width, shot.Height, tex.Width, tex.Height);
         // #931: icon templates are no longer embedded — they're loaded from the
         // asset-extractor cache dir. For this fixture-gated replay the cache (a dev
